liver_system/domo/app/requirements.txt
scikit_image==0.15.0
numpy==1.16.5
torchvision==0.5.0
Flask_Dropzone==1.5.4
nibabel==3.1.0
alembic==1.4.2
Flask_RESTful==0.3.8
Flask_SQLAlchemy==2.4.1
Flask_WTF==0.14.3
Flask_Bcrypt==0.7.1
Flask_Bootstrap==3.3.7.1
Flask==1.1.2
WTForms==2.2.1
SQLAlchemy==1.3.16
matplotlib==3.1.1
Werkzeug==0.16.0
torch==1.4.0
Flask_Login==0.5.0
pydicom==1.4.2
Flask_Migrate==2.5.3
Flask_Script==2.0.6
Pillow==7.1.2
skimage==0.0
